backend/pyproject.toml
[project]
name = "kinship-backend"
version = "1.0.0"
description = "Face Kinship Verification API — FastAPI + AdaFace + RetinaFace"
requires-python = ">=3.10,<3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "numpy>=1.24.0,<2.0.0",
    "Pillow>=10.0.0",
    "huggingface_hub>=0.20.0",
    "safetensors>=0.4.0",
    "omegaconf>=2.3.0",
    "fvcore>=0.1.5",
    "facenet-pytorch>=2.5.0",
    "scikit-learn>=1.3.0",
    "opencv-python-headless>=4.8.0",
]
